Benchmark

+++ experiments/fibonacci/fib.h
@@ -1,5 +1,6 @@
 #ifndef FIB_H
 #define FIB_H
+#include <stdint.h>
 
 // Computing Fibonacci number by using
 // F(2n) = F(n)^2 + F(n+1)^2 and F(2n+1) = 2F(n)F(n+1) + F(n+1)^2
@@ -7,7 +8,7 @@
 // An interative program to calculate fibonacci numbers in O(log n) arithmetic
 // operations, Information Processing Letters Volume 7, Issue 6,
 // October 1978, Pages 299-303 
-unsigned doubling(unsigned n);
+int64_t doubling(int n);
 
 // This function computes the n-th fibonacci number, where n should be a
 // non-negative number
@@ -15,12 +16,12 @@ unsigned doubling(unsigned n);
 // Tumble, a fast simple iteration algorithm for Fibonacci,
 // Information Processing Letters Volume 89, Issue 4, 28 February 2004,
 // Pages 187-189
-unsigned tumble(unsigned n);
+int64_t tumble(int n);
 
 // Ordinary way of computing fibonacci number
-unsigned iterative(unsigned n);
+int64_t iterative(int n);
 
 // Q-matrix method
-unsigned qmatrix(unsigned n);
+int64_t qmatrix(int n);
 
 #endif
